Move file hashing utils to hashing.py
